+++ lib/src/odbc_blocking/init.dart
@@ -1,9 +1,9 @@
-part of './base.dart';
+part of 'base.dart';
 
-extension on DartOdbc {
+extension on DartOdbcBlockingClient {
   void _initialize() {
     final pHEnv = calloc<SQLHANDLE>();
-    tryOdbc(
+    _tryOdbc(
       _sql.SQLAllocHandle(SQL_HANDLE_ENV, nullptr, pHEnv),
       operationType: SQL_HANDLE_ENV,
       handle: pHEnv.value,
@@ -14,7 +14,7 @@ extension on DartOdbc {
     );
     _hEnv = pHEnv.value;
 
-    tryOdbc(
+    _tryOdbc(
       _sql.SQLSetEnvAttr(
         _hEnv,
         SQL_ATTR_ODBC_VERSION,